tests/requirements.txt
aioxmpp==0.11.0
pip-chill==1.0.0
pytest-asyncio==0.14.0
